Beginnings of Profile Component
I will further develop with appropriate text fields and such. I realize
we may have to do some refactoring later if the style needs to be
adjusted or Modified. I will soon have some information on what I did
and why ( although the why may not carry with it much reason other than
it was the only way that I could accomplish the task). Please give me
feedback, look it over, let me know if it works and what we are feeling.
I will do the same for the load profile unless there are any ideas you
guys might have that might be a better fit.