JSON Request
Can now call JSON data from our API. There is some bug present, where
the geoLocation sometimes reads in as an Array and sometimes as an
Object. I believe it's because an Array of size 1 is a special case from
where the server converts the XML database to JSON. Will look into work
around.