More additions to save and load
I have added more functionality for navigation and the appropriate
fields and structures for each screen. I am going to research the most
INefficient ways to save data internally for an app and avoid those the
best that I can. Hopefully after tomorrow we should have a proper load
and save system running.