Bug Fix and UI Change
Created a new   "item_list.xml" activity that functions like the old
one, but can now be rendered in the viewer which allows us easier
modification.

Our app now has color! The current scheme fits a spectrum of oranges and
blues. The darker blues border the outline of the app, drawing more
natural attention to the lighter blue area. The oranges are a little
sharp for now, but should draw attention to the important areas of the
app. Most fonts are now bold to stand out against the lighter colors.
The list items themselves are in yet a lighter blue which should be soft
on the eyes, however, stands out to the user.

Currently the entries use a red header and green fields. Many people I
showcased the app to disliked the red because it felt like it should be
reserved for errors.

Located the bug that was causing our list to sometimes display blank.
Occasionally our JSON request isn't null, isn't returning an error, but
is occasionally just returning a null array/object and throwing an
exception. We now have a handler object in place that will attempt to
retry the call after 4 seconds, and make 3 calls. There is a location
for where our "failure" case should go, but as of now it serves error
testing.

The checkboxes are now radio buttons and located in a radio control so
that only one may be selected at once. Will work on adding sorts next
time based on these selections.