Features
Added sorts for:
Distance/Price/Value

Waiting on client team to determine how to calculate value, but
commented where this calculation needs to go within the JSON code.

Tucked the JSON call into a function. Generate Button now calls the JSON
code if MPG field is not empty, otherwise it calls a function to handle
an empty MPG field.

Changed JSON retries to 4, up from 3, and lowered the time from 4 to 2.5
seconds.

This may be further tweaked, so far it seems stable but we definitely
need to try to break this during bug testing.
Values may shift towards 6 attempts and 2 seconds if stable.

Adjusted list_item such that it appears more natural to the eye, and now
displays which type of fuel the user is looking up.