profileScreen save functionality
Created code for saving profile information to an internal file. This
will create the file if it does not already exist, and append all
entries to the end of the file. The file entries will be a single string
of profileName concatenated with MPG, delimited by a comma.

Exception throwing has been included in all needed function calls.
There is a snippet of code for testing purposes, that should display a
toast with file information, after saving and exiting the
activity_profile_screen.

BUGS: The toast for testing file contents is displaying numbers, rather
than the String that should be saving to the file.