Implemented Load Default Profile Functionality
Implemented functionality of setting a default profile from the Load
Profile screen. This functionality saves the profile information
selected by the user in the Load Profile screen’s ListView to a
Default.txt file. This file is read and automatically applied to the
appropriate text fields in activity_main when the app opens (meaning
the user won’t have to select a profile each time the app is closed and
re-opened - the app will “remember” the user’s latest default profile),
and when activity_main resumes from another activity (so as to update
any time the user might select a new default profile).
Note: Fuel Type data is not included with the default profile
information at this time, only Profile Name and MPG.

Cleaned up profileScreen.java and started some code for Fuel Type
selection in saved profiles (code is commented out, since it is
incomplete at this time).